Reconcile improvements, Optional PVC deletion, envvar configuration, and Status Conditions
Make common create/delete functions for the API objects, to simplify the
Reconcile logic, and better separate functionality. Improves the
readability of the Reconcile logic.

Each Wordpress instance now takes an optional `retainVolumes = true`
setting. If set to **true**, the PVCs created by the Operator for the
deployment instance will not be deleted when the rest of the instance is
removed. By default, the PVCs will be deleted.

Allow environment variables in the operator deployment to configure the
images to use, PVC sizing, secret key naming.

Add and expose some 'Status' from the API for the Wordpress instance.
The Status Conditions expose the state of the API objects, and whether
the operator has created them. This is an example usage of Status and
Conditions that could be expanded to monitor individual object states.